Fixed constructor instance variable definer
Necessary to reduce a Reek violation where the original
`#define_variables` method didn't depend on instance state. It was
only a utility method so is better to extract it to a class method to
more clearly distinguish it. Since the class method is very specific to
the constructor, it's been made private since it should have little
utility outside this object.