Added deprecation warning for infused names method
Due to the introduction of the link:https://alchemists.io/projects/containable[Containable] gem, use of infused names terminology no longer makes sense since keys are used for the container so this is important that this gem uses that terminology as well.

This is a revert of the earlier commit: 3ac09557dc06 (Added infused names and deprecation warning for infused keys).

Milestone: minor